refactor(transport): split transport.py into a focused package
Move the four logical concerns of the old 1.3kloc transport.py into
their own modules:

  goggles/_core/transport/
      __init__.py    # re-exports the flat surface tests/routing rely on
      _frames.py     # wire format, env knobs, shm helpers
      _endpoints.py  # _Endpoint protocol + Unix/Windows + selector
      _protocol.py   # Transport Protocol
      _local.py      # LocalTransport + _SendItem + _SENTINEL

`from goggles._core.transport import X` keeps working for every
public and underscore-prefixed symbol the test suite touched
(LocalTransport, Transport, _pack_small_frame, _unpack_small,
_HEADER_FMT, _IS_WINDOWS, _MSG_SMALL, _SHM_NAME_PREFIX,
_default_socket_path, _next_shm_name, _reap_orphan_shm,
_try_unlink_shm, _user_tag, ...). No behaviour change; git tracks
transport.py → _local.py as a rename so blame still resolves on
the LocalTransport body.

Closes #160.
