+++ goggles/_core/transport/_local.py
@@ -1,618 +1,64 @@
-"""LocalTransport: cross-platform local-machine transport for logging.
+"""``LocalTransport``: cross-platform same-machine transport.
 
-The first process to bind the configured socket path becomes the host (owns
-the :class:`EventBus` and dispatches events to attached handlers); subsequent
-processes on the same machine connect as clients.
+The first process to bind the configured socket path becomes the host
+(owns the :class:`EventBus` and dispatches events to attached handlers);
+subsequent processes on the same machine connect as clients.
 
 On Unix (Linux, macOS) the transport uses ``AF_UNIX`` streams with the
 socket file protected at ``0o600``. On Windows, ``AF_UNIX`` is unreliable
 across Python versions, so the transport binds a TCP loopback socket on
-``127.0.0.1`` and writes the chosen port to a sidecar discovery file at the
-same logical ``socket_path``. Both paths share the same framing protocol.
+``127.0.0.1`` and writes the chosen port to a sidecar discovery file.
 
-Payloads above :data:`_DEFAULT_SHM_THRESHOLD` bytes take a zero-copy
-shared-memory side-channel: the client writes the numpy buffer into a
-``multiprocessing.shared_memory.SharedMemory`` block and sends only metadata
-over the socket; the host maps the same block, copies the view out, and
-unlinks the block. The client tracks outstanding block names and unlinks
-any that never made it to the host (e.g. because shutdown dropped them).
+Payloads above ``shm_threshold`` bytes take a zero-copy shared-memory
+side-channel: the client writes the numpy buffer into a
+``multiprocessing.shared_memory.SharedMemory`` block and sends only
+metadata over the socket; the host maps the same block, copies the
+view out, and unlinks the block.
 """
 
 from __future__ import annotations
 
 import errno
 import logging
-import os
 import pickle
 import queue
-import secrets
 import socket
 import struct
-import sys
-import tempfile
 import threading
 import time
 from multiprocessing import shared_memory
-from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
 from goggles.types import Event
 
+from ._endpoints import _endpoint
+from ._frames import (
+    _HEADER_FMT,
+    _HEADER_SIZE,
+    _IS_WINDOWS,
+    _MSG_ATTACH,
+    _MSG_BYE,
+    _MSG_DETACH,
+    _MSG_LARGE,
+    _MSG_SMALL,
+    _default_shm_threshold,
+    _default_socket_path,
+    _next_shm_name,
+    _pack_large,
+    _pack_small_frame,
+    _reap_orphan_shm,
+    _recvall,
+    _try_unlink_shm,
+    _unpack_large,
+    _unpack_small,
+)
+
 if TYPE_CHECKING:
     from goggles import EventBus
 
 _log = logging.getLogger(__name__)
-
-
-# --- Framing ---------------------------------------------------------------
-
-# Every message on the wire is prefixed with a 1-byte kind and a 4-byte
-# big-endian length. The payload format depends on kind.
-_MSG_SMALL = 1  # inline pickle protocol 5 with out-of-band buffers
-_MSG_LARGE = 2  # shared-memory side-channel; payload is pickled metadata
-_MSG_ATTACH = 3
-_MSG_DETACH = 4
-_MSG_BYE = 5
-
-_HEADER_FMT = "!BI"
-_HEADER_SIZE = struct.calcsize(_HEADER_FMT)
-
-_DEFAULT_SHM_THRESHOLD = 65536
-
-_IS_WINDOWS = sys.platform == "win32"
-
-
-def _user_tag() -> str:
-    """Portable per-user identifier for default socket-path naming.
-
-    Returns:
-        A short string unique per local user.
-    """
-    getuid = getattr(os, "getuid", None)
-    if getuid is not None:
-        return str(getuid())
-    return os.environ.get("USERNAME") or os.environ.get("USER") or "default"
-
-
-def _default_socket_path() -> str:
-    """Default socket path, overridable via ``GOGGLES_SOCKET``.
-
-    On Unix this is the Unix-domain-socket path. On Windows it is the
-    sidecar discovery file that records the TCP port the host is listening
-    on; the file itself is not a socket.
-
-    Returns:
-        Absolute filesystem path.
-    """
-    override = os.getenv("GOGGLES_SOCKET")
-    if override:
-        return override
-    runtime_dir = os.getenv("XDG_RUNTIME_DIR") or tempfile.gettempdir()
-    return os.path.join(runtime_dir, f"goggles-{_user_tag()}.sock")
-
-
-def _default_shm_threshold() -> int:
-    """Default shared-memory threshold in bytes.
-
-    Returns:
-        Minimum payload size (bytes) that triggers the shm side-channel.
-    """
-    raw = os.getenv("GOGGLES_SHM_THRESHOLD")
-    if raw is None:
-        return _DEFAULT_SHM_THRESHOLD
-    try:
-        return max(0, int(raw))
-    except ValueError:
-        _log.warning(
-            "Invalid GOGGLES_SHM_THRESHOLD=%r; using default %d",
-            raw,
-            _DEFAULT_SHM_THRESHOLD,
-        )
-        return _DEFAULT_SHM_THRESHOLD
-
-
-def _recvall(sock: socket.socket, n: int) -> bytes | None:
-    """Read exactly ``n`` bytes from ``sock``.
-
-    Args:
-        sock: Connected stream socket.
-        n: Number of bytes to read.
-
-    Returns:
-        The read bytes, or None if the peer closed the connection.
-    """
-    buf = bytearray()
-    while len(buf) < n:
-        try:
-            chunk = sock.recv(n - len(buf))
-        except (ConnectionResetError, OSError):
-            return None
-        if not chunk:
-            return None
-        buf.extend(chunk)
-    return bytes(buf)
-
-
-def _pack_small_frame(event: Event) -> bytearray:
-    """Build the full MSG_SMALL wire frame in a single allocation.
-
-    The total frame size (5-byte header + body) is computed up front,
-    the bytearray is sized once, and every part (header, length
-    prefixes, main pickle, each out-of-band buffer) is written via
-    ``struct.pack_into`` / slice-assign — one memcpy per part. The send
-    loop then hands this bytearray straight to ``sendall``, so there is
-    no intermediate ``bytes(...)`` cast and no header concatenation.
-
-    Args:
-        event: Event to serialize.
-
-    Returns:
-        A complete wire frame (header + body), ready for ``sendall``.
-    """
-    buffers: list[pickle.PickleBuffer] = []
-    main = pickle.dumps(event, protocol=5, buffer_callback=buffers.append)
-    raws = [buf.raw() for buf in buffers]
-
-    body_len = 4 + len(main) + 4 + sum(4 + len(r) for r in raws)
-    out = bytearray(_HEADER_SIZE + body_len)
-
-    struct.pack_into(_HEADER_FMT, out, 0, _MSG_SMALL, body_len)
-    offset = _HEADER_SIZE
-
-    struct.pack_into("!I", out, offset, len(main))
-    offset += 4
-    out[offset : offset + len(main)] = main
-    offset += len(main)
-
-    struct.pack_into("!I", out, offset, len(buffers))
-    offset += 4
-
-    for raw in raws:
-        n = len(raw)
-        struct.pack_into("!I", out, offset, n)
-        offset += 4
-        out[offset : offset + n] = raw
-        offset += n
-
-    return out
-
-
-def _unpack_small(payload: bytes) -> Event:
-    """Reverse the body half of :func:`_pack_small_frame`.
-
-    The reader strips the 5-byte header and passes the remaining body
-    to this function.
-
-    Args:
-        payload: Frame body (everything after ``_HEADER_SIZE``).
-
-    Returns:
-        The deserialized Event.
-    """
-    offset = 0
-    (main_len,) = struct.unpack_from("!I", payload, offset)
-    offset += 4
-    main = payload[offset : offset + main_len]
-    offset += main_len
-    (num_bufs,) = struct.unpack_from("!I", payload, offset)
-    offset += 4
-    buffers: list[bytes] = []
-    for _ in range(num_bufs):
-        (blen,) = struct.unpack_from("!I", payload, offset)
-        offset += 4
-        buffers.append(payload[offset : offset + blen])
-        offset += blen
-    return pickle.loads(main, buffers=buffers)
-
-
-def _pack_large(event: Event, shm_name: str) -> bytes:
-    """Serialize a LARGE-mode event.
-
-    The numpy payload has already been copied into shared memory named
-    ``shm_name``; the wire payload carries metadata needed to reconstruct
-    the array on the host side.
-
-    Args:
-        event: Event whose ``payload`` is a numpy ndarray.
-        shm_name: Name of the shared memory block holding the array bytes.
-
-    Returns:
-        Pickled metadata + a stripped event (payload replaced by None).
-
-    Raises:
-        TypeError: If ``event.payload`` is not a ``numpy.ndarray``.
-    """
-    arr = event.payload
-    if not isinstance(arr, np.ndarray):
-        raise TypeError(
-            "LARGE path requires numpy.ndarray payload, got "
-            f"{type(arr).__name__}"
-        )
-    stripped = Event(
-        kind=event.kind,
-        scope=event.scope,
-        payload=None,
-        filepath=event.filepath,
-        lineno=event.lineno,
-        level=event.level,
-        step=event.step,
-        time=event.time,
-        extra=event.extra,
-    )
-    meta: dict[str, Any] = {
-        "shm_name": shm_name,
-        "dtype": str(arr.dtype),
-        "shape": tuple(arr.shape),
-        "nbytes": int(arr.nbytes),
-        "event": stripped,
-    }
-    return pickle.dumps(meta, protocol=5)
-
-
-def _unpack_large(payload: bytes) -> Event:
-    """Reconstruct an Event from a LARGE frame and unlink the named shm.
-
-    Args:
-        payload: Pickled metadata produced by :func:`_pack_large`.
-
-    Returns:
-        The reconstructed Event with its numpy payload materialized.
-    """
-    meta = pickle.loads(payload)
-    shm_name: str = meta["shm_name"]
-    dtype = np.dtype(meta["dtype"])
-    shape: tuple[int, ...] = tuple(meta["shape"])
-    stripped: Event = meta["event"]
-
-    shm = shared_memory.SharedMemory(name=shm_name)
-    try:
-        view = np.ndarray(shape, dtype=dtype, buffer=shm.buf)
-        arr = np.array(view, copy=True)
-    finally:
-        shm.close()
-        _try_unlink_shm(shm_name)
-    return Event(
-        kind=stripped.kind,
-        scope=stripped.scope,
-        payload=arr,
-        filepath=stripped.filepath,
-        lineno=stripped.lineno,
-        level=stripped.level,
-        step=stripped.step,
-        time=stripped.time,
-        extra=stripped.extra,
-    )
-
-
-_SHM_NAME_PREFIX = "goggles_"
-_SHM_REAP_AGE_S = 300.0
-_LINUX_SHM_DIR = "/dev/shm"
-
-
-def _next_shm_name() -> str:
-    """Return a uniquely-prefixed shared-memory name for this process.
-
-    The ``goggles_`` prefix lets the host opportunistically reap blocks
-    that survived a crash without the consumer ever processing them.
-
-    Returns:
-        A name suitable for ``SharedMemory(create=True, name=...)``.
-    """
-    return f"{_SHM_NAME_PREFIX}{os.getpid()}_{secrets.token_hex(8)}"
-
-
-def _reap_orphan_shm(max_age_s: float = _SHM_REAP_AGE_S) -> int:
-    """Best-effort sweep of stale ``goggles_*`` shared-memory blocks.
-
-    The :mod:`multiprocessing.shared_memory` resource tracker covers
-    in-process leaks, but a host that crashes between receiving a
-    LARGE frame and unlinking the named segment leaves the block
-    behind. This sweep runs at host startup and unlinks any
-    ``goggles_*`` segment older than ``max_age_s`` seconds. Linux only
-    (``/dev/shm`` is the visible mount); a no-op elsewhere.
-
-    Args:
-        max_age_s: Reap blocks whose mtime is older than this many
-            seconds. The default leaves enough headroom that a busy
-            client's just-allocated segments are never touched.
-
-    Returns:
-        Count of segments unlinked (zero on non-Linux or empty dir).
-    """
-    if not os.path.isdir(_LINUX_SHM_DIR):
-        return 0
-    try:
-        names = os.listdir(_LINUX_SHM_DIR)
-    except OSError:
-        return 0
-    cutoff = time.time() - max_age_s
-    reaped = 0
-    for name in names:
-        if not name.startswith(_SHM_NAME_PREFIX):
-            continue
-        path = os.path.join(_LINUX_SHM_DIR, name)
-        try:
-            if os.path.getmtime(path) >= cutoff:
-                continue
-        except OSError:
-            continue
-        try:
-            os.unlink(path)
-            reaped += 1
-        except OSError:
-            pass
-    return reaped
-
-
-def _try_unlink_shm(name: str) -> None:
-    """Best-effort unlink of a named shm block, tolerant to already-gone.
-
-    Args:
-        name: Name of the shared-memory block to remove.
-    """
-    try:
-        shm = shared_memory.SharedMemory(name=name)
-    except FileNotFoundError:
-        return
-    except OSError:
-        return
-    try:
-        shm.close()
-    except OSError:
-        pass
-    try:
-        shm.unlink()
-    except (FileNotFoundError, OSError):
-        pass
-
-
-# --- Cross-platform endpoint ----------------------------------------------
-
-
-class _Endpoint(Protocol):
-    """Platform abstraction for binding / connecting to a logical path."""
-
-    @staticmethod
-    def connect(path: str, timeout: float = 1.0) -> socket.socket | None:
-        """Attempt to connect to a host at ``path``.
-
-        Args:
-            path: Logical endpoint identifier.
-            timeout: Connect timeout in seconds.
-
-        Returns:
-            Connected stream socket, or None if no host is listening.
-        """
-        ...
-
-    @staticmethod
-    def bind(path: str) -> socket.socket:
-        """Bind and listen at ``path``.
-
-        Args:
-            path: Logical endpoint identifier.
-
-        Returns:
-            The bound server socket.
-        """
-        ...
-
-    @staticmethod
-    def cleanup(path: str) -> None:
-        """Remove filesystem artifacts associated with ``path``.
-
-        Best-effort; callers should not rely on failure modes.
-
-        Args:
-            path: Logical endpoint identifier.
-        """
-        ...
-
-    @staticmethod
-    def accept_address_hint() -> str:
-        """Return a human-readable label for logs.
-
-        Returns:
-            A short string identifying the endpoint family.
-        """
-        ...
-
-
-class _UnixEndpoint:
-    """AF_UNIX-based endpoint for Linux and macOS."""
-
-    @staticmethod
-    def connect(path: str, timeout: float = 1.0) -> socket.socket | None:
-        if not os.path.exists(path):
-            return None
-        sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-        sock.settimeout(timeout)
-        try:
-            sock.connect(path)
-        except OSError:
-            sock.close()
-            return None
-        sock.settimeout(None)
-        return sock
-
-    @staticmethod
-    def bind(path: str) -> socket.socket:
-        parent = os.path.dirname(path)
-        if parent:
-            os.makedirs(parent, exist_ok=True)
-            try:
-                os.chmod(parent, 0o700)
-            except OSError:
-                # Parent may be a shared dir (e.g. /tmp); don't fail here.
-                pass
-
-        server = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-        old_umask = os.umask(0o077)
-        try:
-            server.bind(path)
-        finally:
-            os.umask(old_umask)
-        try:
-            os.chmod(path, 0o600)
-        except OSError:
-            pass
-        server.listen(64)
-        return server
-
-    @staticmethod
-    def cleanup(path: str) -> None:
-        try:
-            os.unlink(path)
-        except (FileNotFoundError, IsADirectoryError):
-            pass
-        except OSError:
-            _log.exception("Failed to unlink socket file at %s", path)
-
-    @staticmethod
-    def accept_address_hint() -> str:
-        return "AF_UNIX"
-
-
-class _WindowsEndpoint:
-    """TCP loopback endpoint for Windows.
-
-    The "socket path" becomes a sidecar file recording the port the host
-    chose; clients read it to find the host.
-    """
-
-    _LOOPBACK = "127.0.0.1"
-
-    @staticmethod
-    def _read_port(path: str) -> int | None:
-        try:
-            with open(path, encoding="utf-8") as f:
-                data = f.read().strip()
-        except OSError:
-            return None
-        try:
-            return int(data)
-        except ValueError:
-            return None
-
-    @staticmethod
-    def connect(path: str, timeout: float = 1.0) -> socket.socket | None:
-        port = _WindowsEndpoint._read_port(path)
-        if port is None:
-            return None
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        sock.settimeout(timeout)
-        try:
-            sock.connect((_WindowsEndpoint._LOOPBACK, port))
-        except OSError:
-            sock.close()
-            return None
-        sock.settimeout(None)
-        sock.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
-        return sock
-
-    @staticmethod
-    def bind(path: str) -> socket.socket:
-        parent = os.path.dirname(path)
-        if parent:
-            os.makedirs(parent, exist_ok=True)
-        server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        server.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        server.bind((_WindowsEndpoint._LOOPBACK, 0))
-        server.listen(64)
-        port = server.getsockname()[1]
-        # Write port atomically: write to tmp, then rename.
-        tmp = path + ".tmp"
-        with open(tmp, "w", encoding="utf-8") as f:
-            f.write(str(port))
-        os.replace(tmp, path)
-        return server
-
-    @staticmethod
-    def cleanup(path: str) -> None:
-        try:
-            os.unlink(path)
-        except (FileNotFoundError, IsADirectoryError):
-            pass
-        except OSError:
-            _log.exception("Failed to remove discovery file at %s", path)
-
-    @staticmethod
-    def accept_address_hint() -> str:
-        return "TCP loopback"
-
-
-def _endpoint() -> type[_Endpoint]:
-    if _IS_WINDOWS:
-        return _WindowsEndpoint  # type: ignore[return-value]
-    return _UnixEndpoint  # type: ignore[return-value]
-
-
-# --- Transport Protocol ----------------------------------------------------
-
-
-@runtime_checkable
-class Transport(Protocol):
-    """Contract for routing Goggles events to the EventBus."""
-
-    @property
-    def is_running(self) -> bool:
-        """Whether the transport is accepting new events.
-
-        Returns:
-            True while the transport is live, False after shutdown or an
-            unrecoverable send failure.
-        """
-        ...
-
-    def emit(self, event: Event) -> None:
-        """Route an event asynchronously (fire-and-forget).
-
-        Args:
-            event: Event to route.
-        """
-        ...
-
-    def emit_sync(self, event: Event) -> None:
-        """Route an event, blocking until it has reached the bus.
-
-        In client mode this is best-effort: there is no cross-process ack,
-        so it behaves like :meth:`emit` plus a local queue flush.
-
-        Args:
-            event: Event to route.
-        """
-        ...
-
-    def attach(self, handlers: list[dict], scopes: list[str]) -> None:
-        """Attach handlers under the given scopes.
-
-        Args:
-            handlers: Serialized handler dicts (see ``Handler.to_dict``).
-            scopes: Scopes under which to attach.
-        """
-        ...
-
-    def detach(self, handler_name: str, scope: str) -> None:
-        """Detach a handler from the given scope.
-
-        Args:
-            handler_name: Name of the handler to detach.
-            scope: Scope to detach from.
-        """
-        ...
-
-    def shutdown(self, timeout: float | None = None) -> None:
-        """Flush pending events and release resources.
-
-        Args:
-            timeout: Max seconds per background thread for clean exit.
-                ``None`` means wait indefinitely.
-        """
-        ...
-
-
-# --- LocalTransport --------------------------------------------------------
 
 
 _SENTINEL = object()  # "stop draining / sending" marker
